note-digitizer/requirements.txt
google-generativeai>=0.8.0
watchdog>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
pillow-heif>=0.16.0
